Bind部分で個別に分けていたインタフェースを統合
Viewからのアクセス範囲をBindのみに絞るなら後で再度分割する
