+++ learn-flux-cs/Flux/Stores/Store.cs
@@ -12,7 +12,7 @@ namespace LearnFlux.Flux.Stores;
 public abstract class Store( IDispatcher dispatcher ) : IStore
 {
     protected readonly IDispatcher dispatcher = dispatcher;
-    private readonly StoreBinder storeBinder = new();
+    private readonly Dictionary<object, List<object>> bindings = new();
 
     ///
     /// <inheritdoc />
@@ -39,12 +39,65 @@ public abstract class Store( IDispatcher dispatcher ) : IStore
     ///
     /// <inheritdoc />
     ///
-    public virtual IDisposable Bind<TActionType, TPayload>( TActionType actionType, Func<TPayload, Task> callback )
-        => storeBinder.Bind( actionType, callback );
+    public IDisposable Bind<TActionType, TPayload>( TActionType actionType, Func<TPayload, Task> callback )
+    {
+        _ = actionType ?? throw new ArgumentNullException( nameof( actionType ) );
+        _ = callback ?? throw new ArgumentNullException( nameof( callback ) );
+
+        if( !bindings.TryGetValue( actionType, out var callbacks ) )
+        {
+            callbacks              = [];
+            bindings[ actionType ] = callbacks;
+        }
+
+        if( callbacks.Contains( callback ) )
+        {
+            throw new InvalidOperationException( $"{actionType} is already bound" );
+        }
+
+        callbacks.Add( callback );
+
+        return new BindToken( this, actionType, callback );
+    }
 
     ///
     /// <inheritdoc />
     ///
     public IEnumerable<Func<TPayload, Task>> CallbacksOf<TActionType, TPayload>( TActionType actionType )
-        => storeBinder.CallbacksOf<TActionType, TPayload>( actionType );
+    {
+        _ = actionType ?? throw new ArgumentNullException( nameof( actionType ) );
+
+        if( bindings.TryGetValue( actionType, out var callback ) )
+        {
+            var result = new List<Func<TPayload, Task>>();
+
+            foreach( var x in callback )
+            {
+                result.Add( (Func<TPayload, Task>)x );
+            }
+
+            return result;
+        }
+
+        return Array.Empty<Func<TPayload, Task>>();
+    }
+
+    private class BindToken( Store store, object actionType, object callback ) : IDisposable
+    {
+        private readonly Store store = store;
+        private readonly object actionType = actionType;
+        private readonly object callback = callback;
+
+        public void Dispose()
+        {
+            _ = actionType ?? throw new InvalidOperationException( $"{nameof( actionType )} is null" );
+
+            var bindings = store.bindings;
+
+            if( bindings.TryGetValue( actionType, out var callbacks ) )
+            {
+                callbacks.Remove( callback );
+            }
+        }
+    }
 }
